Introduce initial XR based example
This is based upon the 05a_materials example so that we can load
in a glTF 2 file and render it efficiently.

The XR integration uses the multiview rendering approach. This works
fine on nVidia GPUs but seems to still have issues with AMD GPUs. I suspect
there is a missing synchronization point somewhere but requires more
investigation.

Left thumbstick can be used to position model in xy-plane and the right
thumbstick can be used to adjust the model scale. More controls will
be added over time.