Introduce simple Slint interop example
We use KDGpu to render to an offscreen texture, which we export via
OpaqueFd/OpaqueWin32. We then import this to an OpenGL texture, which we
use to construct a Slint Image using
Image::create_from_borrowed_gl_2d_rgb_texture.

There's no copying involved. We're sharing a single texture in GPU
memory.

This assumes Slint is using the Winit-FemtoVG backend.

This example is missing double buffering and proper synchronization.

Change-Id: I38c5d383e46f316a827739f4b8e1a5955e6ba9ae
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu-examples/+/152763
Reviewed-by: Paul Lemire <paul.lemire@kdab.com>