ffmpeg_video: use dedicated thread for decoding
- Move decoding to a dedicated thread
- Decoder now fills a queue of up to 5 decoded frames (blocking if no room in queue)
- Rendering now asks queue for the last decoded frame
- Based on frame timestamp we know if the frame ought to be
  uploaded immediately or if we are better duplicating the current frame

Throttling of the decoding thread is done by the speed at which the
renderer takes and upload frames.
FrameRate is computed by using the delta between last time we uploaded a frame
and current time. We compare that delta against the time at which we expect
to upload the next frame.

Change-Id: I79b6cace23e5bd8230f45176293cc2d437b75753
Reviewed-on: https://codereview.kdab.com/c/kdab/kdgpu-examples/+/153057
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>